(PC-12846)[PRO] refactor(imageEditor): translate to TypeScript
