+++ pro/package.json
@@ -114,6 +114,7 @@
     "@types/lodash.isequal": "^4.5.5",
     "@types/node": "^16.7.10",
     "@types/react": "^17.0.19",
+    "@types/react-avatar-editor": "^10.3.6",
     "@types/react-datepicker": "^4.3.2",
     "@types/react-dom": "^17.0.9",
     "@types/react-redux": "^7.1.18",
